requirements.txt
openai==1.3.7
aws_requests_auth==0.4.3
requests==2.31.0
boto3==1.33.7
pytest-mock==3.12.0
websocket-client==1.7.0
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
deepdiff==6.7.1